web-animations/interfaces/KeyframeEffect/processing-a-keyframes-argument.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>KeyframeEffectReadOnly constructor tests</title>
<link rel="help" href="https://w3c.github.io/web-animations/#processing-a-keyframes-argument">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<script src="../../resources/keyframe-utils.js"></script>
<body>
<div id="log"></div>
<div id="target"></div>
<script>
'use strict';

// Test the "process a keyframe-like object" procedure.
//
// This file only tests the KeyframeEffectReadOnly constructor since it is
// assumed that the implementation of the KeyframeEffect constructor,
// Animatable.animate() method, and KeyframeEffect.setKeyframes() method will
// all share common machinery and it is not necessary to test each method.

// Test that only animatable properties are accessed

var gNonAnimatableProps = [
  'animation', // Shorthands where all the longhand sub-properties are not
               // animatable, are also not animatable.
  'animationDelay',
  'animationDirection',
  'animationDuration',
  'animationFillMode',
  'animationIterationCount',
  'animationName',
  'animationPlayState',
  'animationTimingFunction',
  'transition',
  'transitionDelay',
  'transitionDuration',
  'transitionProperty',
  'transitionTimingFunction',
  'display',
  'unsupportedProperty',
];

function TestKeyframe(testProp) {
  var _propAccessCount = 0;

  Object.defineProperty(this, testProp, {
    get: function() { _propAccessCount++; },
    enumerable: true
  });

  Object.defineProperty(this, 'propAccessCount', {
    get: function() { return _propAccessCount; }
  });
}

function GetTestKeyframeSequence(testProp) {
  return [ new TestKeyframe(testProp) ]
}

gNonAnimatableProps.forEach(function(prop) {
  test(function(t) {
    var testKeyframe = new TestKeyframe(prop);

    new KeyframeEffectReadOnly(null, testKeyframe);

    assert_equals(testKeyframe.propAccessCount, 0, 'Accessor not called');
  }, 'non-animatable property \'' + prop + '\' is not accessed when using'
     + ' a property-indexed keyframe object');
});

gNonAnimatableProps.forEach(function(prop) {
  test(function(t) {
    var testKeyframes = GetTestKeyframeSequence(prop);

    new KeyframeEffectReadOnly(null, testKeyframes);

    assert_equals(testKeyframes[0].propAccessCount, 0, 'Accessor not called');
  }, 'non-animatable property \'' + prop + '\' is not accessed when using'
     + ' a keyframe sequence');
});

// Test handling of custom iterable objects.

function createIterable(iterations) {
  return {
    [Symbol.iterator]() {
      var i = 0;
      return {
        next() {
          return iterations[i++];
        },
      };
    },
  };
}

test(() => {
  var effect = new KeyframeEffect(null, createIterable([
    {done: false, value: {left: '100px'}},
    {done: false, value: {left: '300px'}},
    {done: false, value: {left: '200px'}},
    {done: true},
  ]));
  assert_frame_lists_equal(effect.getKeyframes(), [
    {offset: null, computedOffset: 0, easing: 'linear', left: '100px'},
    {offset: null, computedOffset: 0.5, easing: 'linear', left: '300px'},
    {offset: null, computedOffset: 1, easing: 'linear', left: '200px'},
  ]);
}, 'Custom iterator with basic keyframes.');

test(() => {
  var keyframes = createIterable([
    {done: false, value: {left: '100px'}},
    {done: false, value: {left: '300px'}},
    {done: false, value: {left: '200px'}},
    {done: true},
  ]);
  keyframes.easing = 'ease-in-out';
  keyframes.offset = '0.1';
  var effect = new KeyframeEffect(null, keyframes);
  assert_frame_lists_equal(effect.getKeyframes(), [
    {offset: null, computedOffset: 0, easing: 'linear', left: '100px'},
    {offset: null, computedOffset: 0.5, easing: 'linear', left: '300px'},
    {offset: null, computedOffset: 1, easing: 'linear', left: '200px'},
  ]);
}, 'easing and offset are ignored on iterable objects.');

test(() => {
  var effect = new KeyframeEffect(null, createIterable([
    {done: false, value: {left: '100px', top: '200px'}},
    {done: false, value: {left: '300px'}},
    {done: false, value: {left: '200px', top: '100px'}},
    {done: true},
  ]));
  assert_frame_lists_equal(effect.getKeyframes(), [
    {offset: null, computedOffset: 0, easing: 'linear', left: '100px', top: '200px'},
    {offset: null, computedOffset: 0.5, easing: 'linear', left: '300px'},
    {offset: null, computedOffset: 1, easing: 'linear', left: '200px', top: '100px'},
  ]);
}, 'Custom iterator with multiple properties specified.');

test(() => {
  var effect = new KeyframeEffect(null, createIterable([
    {done: false, value: {left: '100px'}},
    {done: false, value: {left: '250px', offset: 0.75}},
    {done: false, value: {left: '200px'}},
    {done: true},
  ]));
  assert_frame_lists_equal(effect.getKeyframes(), [
    {offset: null, computedOffset: 0, easing: 'linear', left: '100px'},
    {offset: 0.75, computedOffset: 0.75, easing: 'linear', left: '250px'},
    {offset: null, computedOffset: 1, easing: 'linear', left: '200px'},
  ]);
}, 'Custom iterator with offset specified.');

test(() => {
  assert_throws({name: 'TypeError'}, function() {
    new KeyframeEffect(null, createIterable([
      {done: false, value: {left: '100px'}},
      {done: false, value: 1234},
      {done: false, value: {left: '200px'}},
      {done: true},
    ]));
  });
}, 'Custom iterator with non object keyframe should throw.');

test(() => {
  var effect = new KeyframeEffect(null, createIterable([
    {done: false, value: {left: ['100px', '200px']}},
    {done: true},
  ]));
  assert_frame_lists_equal(effect.getKeyframes(), [
    {offset: null, computedOffset: 1, easing: 'linear', left: '100px,200px'}
  ]);
}, 'Custom iterator with value list in keyframe should give bizarre string representation of list.');

// FIXME: Test that non-enumerable properties are not accessed

// FIXME: Test that properties are accessed in ascending order by Unicode
//        codepoint
//        (There is an existing test for this in
//        keyframe-effect/constructor.html that should be moved here.)

</script>
